requirements.txt
# Core Django - successfully installed
Django>=4.2.0,<5.0.0
asgiref>=3.6.0,<4.0.0
sqlparse>=0.4.0,<1.0.0
pytz>=2023.3,<2024.0
tzdata>=2023.3,<2024.0

# HTTP and async - successfully installed with pre-compiled wheels
aiohttp>=3.12.0,<4.0.0
aiohttp-retry>=2.9.0,<3.0.0
aiosignal>=1.4.0,<2.0.0
frozenlist>=1.7.0,<2.0.0
multidict>=6.6.0,<7.0.0
yarl>=1.20.0,<2.0.0
aiohappyeyeballs>=2.6.0,<3.0.0
propcache>=0.3.0,<1.0.0

# HTTP requests - successfully installed
requests>=2.32.0,<3.0.0
urllib3>=2.5.0,<3.0.0
charset-normalizer>=3.4.0,<4.0.0
idna>=3.10,<4.0
certifi>=2025.8.0,<2026.0

# Security and crypto - successfully installed with pre-compiled wheels
cryptography>=45.0.0,<46.0.0
cffi>=1.17.0,<2.0.0
pycparser>=2.22,<3.0
pycryptodome>=3.23.0,<4.0.0
PyJWT>=2.10.0,<3.0.0

# Utilities - successfully installed
python-dotenv>=1.1.0,<2.0.0
twilio>=9.7.0,<10.0.0
attrs>=25.3.0,<26.0.0
